Refactoring file structure
As suggested in the FastAPI documentation
https://fastapi.tiangolo.com/tutorial/bigger-applications/#import-fastapi
