+++ exodus_gw/routers/gateway.py
@@ -1,35 +1,18 @@
-from fastapi import Depends
-from fastapi.exception_handlers import http_exception_handler
-from starlette.exceptions import HTTPException as StarletteHTTPException
+from fastapi import APIRouter, Depends
 
-from .app import app
-from .publish import create_publish_id
-from .auth import call_context, CallContext
-from .s3.util import xml_response
+from ..auth import CallContext, call_context
+from ..publish import create_publish_id
+
+router = APIRouter()
 
 
-@app.exception_handler(StarletteHTTPException)
-async def custom_http_exception_handler(request, exc):
-    # Override HTTPException to produce XML error responses for the
-    # given endpoints.
-
-    path = request.scope.get("path")
-
-    if path.startswith("/upload"):
-        return xml_response(
-            "Error", Code=exc.status_code, Message=exc.detail, Endpoint=path
-        )
-
-    return await http_exception_handler(request, exc)
-
-
-@app.get("/healthcheck", tags=["service"])
+@router.get("/healthcheck", tags=["service"])
 def healthcheck():
     """Returns a successful response if the service is running."""
     return {"detail": "exodus-gw is running"}
 
 
-@app.post("/{env}/publish")
+@router.post("/{env}/publish")
 def publish(env: str):
     """WIP: Returns a new, empty publish id"""
     if env not in ["dev", "qa", "stage", "prod"]:
@@ -38,7 +21,7 @@ def publish(env: str):
     return {"detail": "Created Publish Id"}
 
 
-@app.get(
+@router.get(
     "/whoami",
     response_model=CallContext,
     tags=["service"],
