--- exodus_gw/s3/__init__.py
@@ -1,1 +1,0 @@
-from . import api  # noqa
