+++ tests/s3/test_manage_mpu.py
@@ -3,9 +3,10 @@ import textwrap
 import pytest
 
 import mock
-from exodus_gw.s3.api import abort_multipart_upload, multipart_upload
-from exodus_gw.s3.util import xml_response
 from fastapi import HTTPException
+
+from exodus_gw.routers.api import abort_multipart_upload, multipart_upload
+from exodus_gw.s3.util import xml_response
 
 TEST_KEY = "b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c"
