+++ tests/s3/test_upload.py
@@ -1,9 +1,10 @@
-import mock
 import pytest
 
+import mock
 from fastapi import HTTPException
 
-from exodus_gw.s3.api import upload
+from exodus_gw.routers.api import upload
+
 
 TEST_KEY = "b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c"
