+++ tests/test_exception_handler.py
@@ -1,9 +1,9 @@
+import pytest
+
+import mock
 from fastapi import HTTPException
 
-import pytest
-import mock
-
-from exodus_gw.gateway import custom_http_exception_handler
+from exodus_gw.main import custom_http_exception_handler
 
 TEST_KEY = "b5bb9d8014a0f9b1d61e21e796d78dccdf1352f23cd32812f4850b878ae4944c"
